[w] plot and progressbar get data from database

+++ app_calorie/lib/models/totalCal.dart
@@ -3,10 +3,9 @@ import 'package:flutter/material.dart';
 //THis class extends ChangeNotifier. It will act as data repository to be shared thorugh the application.
 class TotalCal extends ChangeNotifier{
 
-  //For simplicity, a product is just a String.
-  double totalCalories = 0;
+  int totalCalories = 0;
 
-  void addCal(double toAdd){
+  void addCal(int toAdd){
     totalCalories += toAdd;
     //Call the notifyListeners() method to alert that something happened.
     notifyListeners();
